Add a method to the mySexCat catalog class to address cases where there are missing lines in the SExtractor ascii catalog header for multiple instances of a given parameter.
For example, if you give a list of photometric apertures in the parameter file, there will only be one "FLUX_APER" line in the header of the output catalog.  This method adds the necessary lines, adding columns like "FLUX_APER2", "FLUX_APER3", to the header to be able to read in the catalog.